First pass of unit tests
